Refactor ComboBoxCustomize & Example
